Peer Review
Fixed error with opening, multiply_click function, and bug for invalid input with value2.
