+++ README.md
@@ -1,2 +1,9 @@
 # Assignment-7
 Assignment #7
+
+Updates:  
+-Moved location of Project files to "Assignment7/WindowsApp/*files*" in order to open the project. (WindowsApp Folder was nested twice "Assignment7/WindowsApp/WindowsApp/*Files*")  
+
+###Updates in Form1.cs
+-Added a second check for res to run for the second value in a nested If/Else Statement.
+-Modified the multiply_click event so it multiplies instead of adds the two values.
